[463737] fix Multitenant ConcurrentModificationException
https://bugs.eclipse.org/bugs/show_bug.cgi?id=463737

For ManagedTypeImpl, if many threads are creating entity managers,
ConcurrentModificationException related to members appears (the second
thread reaches the if (null != this.members) and exits - it will access
members that are in construction, leading to exceptions and invalid
results.

All the tenant client session will clone the classdescriptor from the
client session (same client session). If there are multiple tenant
client sessions started, the notifyReferencingDescriptorsOfIsolation is
called for other entity manager, while referencedDescriptors are added
from the initialize method.

The tests were done with a multitenant-rich data model with 100 parallel
threads doing create an EntityManager (with tenant id), start a
transaction, commit the transaction.
+whitespace fix
# Conflicts:
#	foundation/org.eclipse.persistence.core/src/org/eclipse/persistence/descriptors/ClassDescriptor.java
#	jpa/org.eclipse.persistence.jpa/src/org/eclipse/persistence/internal/jpa/metamodel/ManagedTypeImpl.java